fix logic create segment

+++ src/main/kotlin/io/github/controleagenda/services/SegmentServiceImpl.kt
@@ -14,15 +14,33 @@ class SegmentServiceImpl : SegmentService {
 
     override fun addSegment(id: Long, segment: Segment): Segment {
 
-        println("Esse é o id $id")
+        val allSegments = repository.findAll()
+        var idSequenceOK = false
+        var idSequence: Long = 5
 
-        return repository.save(Segment(id = id, segment.segment))
+        if (allSegments.count() < id && !repository.findById(id).isPresent) {
+            return repository.save(Segment(id, segment.segment))
+        } else {
+            while (!idSequenceOK) {
+                if (repository.findById(idSequence).isPresent) {
+                    idSequence++
+                } else {
+                    idSequenceOK = true
+                }
+            }
+        }
+        return repository.save(Segment(idSequence, segment.segment))
     }
 
     override fun deleteSegment(id: Long) {
-        val segment = repository.findById(id)
-        repository.deleteById(id)
-        println("O usuario ${segment.get().segment} foi deletado com sucesso")
+
+        if (id in 1..6) {
+            throw RuntimeException("Não é possível apagar os valores default")
+        } else {
+            val segment = repository.findById(id)
+            repository.deleteById(id)
+            println("O usuario ${segment.get().segment} foi deletado com sucesso")
+        }
     }
 
     override fun getSegmentById(id: Long): Optional<Segment?> {
@@ -35,9 +53,8 @@ class SegmentServiceImpl : SegmentService {
 
         if (repository.findAll().count() <= 1) {
             return initSegments()
-        } else {
-            return repository.findAll()
         }
+        return repository.findAll()
     }
 
     override fun updateSegments(id: Long, segment: Segment): Segment {
@@ -46,13 +63,11 @@ class SegmentServiceImpl : SegmentService {
 
         if (segment.id!! >= 1) {
             return addSegment(segment.id, segment)
-        } else {
-            return addSegment(id, segment)
         }
+        return addSegment(id, segment)
     }
 
     fun initSegments(): List<Segment> {
-
         repository.save(
             Segment(1, "Academia")
         )
@@ -71,10 +86,8 @@ class SegmentServiceImpl : SegmentService {
         repository.save(
             Segment(6, "Saúde")
         )
-        repository.save(Segment(7, "Outros"))
 
         return repository.findAll() as List<Segment>
-
     }
 
 }
